refactor: integration test common code into internal module
